@@ -101,6 +101,7 @@ internal sealed class YandexDiskHandler: BaseSourceHandler
             await using var stream = await client.GetStreamAsync(uri, cancellationToken).ConfigureAwait(false);
             await handler.FillPipeAsync(stream, writer, cancellationToken).ConfigureAwait(false);
         }
-    }
 
+        public void Dispose() { }
+    }
 }
